refactor(rc-actions): use children in Actions
